refactor(components): add layout folder & update readme
