+++ src/components/layouts/DetailLayout/styles.ts
